- refractor all files naming convention - clean up postgres steps
