+++ load/postgres.py
@@ -1,10 +1,8 @@
 import pandas as pd
+import util as log
 
-from connection.postgres_config import create_engine_config, dispose_engine, check_schema_exist
+from connection.postgres_connection import create_engine_config, dispose_engine
 
-
-# TODO: Drop all table in postgres (WH)
-# def drop_all_table(schema, db_properties):
 
 # Read from db and set it as data frame
 def select_table(sql_query, db_properties):
@@ -19,7 +17,6 @@ def select_table(sql_query, db_properties):
 
 # To insert to database
 def upsert_table(data_frame, table_name, db_properties):
-    check_schema_exist(db_properties)
     engine = create_engine_config(db_properties)
     if engine is not None:
         data_frame.to_sql(table_name, con=engine, if_exists='replace')
@@ -28,8 +25,19 @@ def upsert_table(data_frame, table_name, db_properties):
         dispose_engine(engine)
 
 
-def test_retrieve_postgres(db_properties):
-    check_schema_exist(db_properties)
+def check_schema_exist(db_properties):
+    engine = create_engine_config(db_properties)
+    if engine is not None:
+        # Read
+        sql_query = "SELECT COUNT(schema_name) FROM information_schema.schemata WHERE schema_name = '{}'".format(
+            db_properties.schema)
+        result = engine.execute(sql_query).scalar()
+        if result == 0:
+            # schema does not exist
+            # create schema
+            sql_query = "CREATE SCHEMA IF NOT EXISTS {}".format(db_properties.schema)
+            engine.execute(sql_query)
+            log.message.warning_no_schema(db_properties)
 
 
 # delete all table
